+++ generator.h
@@ -8,13 +8,13 @@
 Creates a new cube object based on a block object.
 material can be NULL
 */
-struct cube cubeFromBlock(struct block block, const int side, struct material* material);
+struct cube cubeFromBlock(struct block block, const unsigned int side, struct material* material);
 
 /*
 Isolated from the main code to maybe in the future enable for a multithreaded version that can do multiple chunks.
 materials can be NULL
 */
-struct cubeModel createCubeModel(struct section* sections, int sectionLen, hashTable* materials, bool yLim, int upLim, int downLim, int side, bool matCheck, int xOff, int zOff);
+struct cubeModel createCubeModel(struct section* sections, int sectionLen, hashTable* materials, bool yLim, int upLim, int downLim, unsigned int side, bool matCheck, int xOff, int zOff);
 
 //Generally speaking i could just pass a whole chunk object from regionParser here, but I like the design of having that be completely separate
 
@@ -22,4 +22,4 @@ struct cubeModel createCubeModel(struct section* sections, int sectionLen, hashT
 Generates a model object from binary nbt data.
 materials and objects can be NULL.
 */
-model generateFromNbt(unsigned char* data, long dataSize, hashTable* materials, hashTable* objects, bool yLim, int upLim, int downLim, bool b, bool f, int side, int chunkX, int chunkZ);
+model generateFromNbt(unsigned char* data, int dataSize, hashTable* materials, hashTable* objects, bool yLim, int upLim, int downLim, bool b, bool f, unsigned int side, int chunkX, int chunkZ);
